Add files via upload
Used Car Price Prediction System is a web and AI based software which allows users to predict the price of the used cars by asking few questions about the features of the car. The purpose of the system is to predict the price of the used cars according to the market. In existing systems someone who wants to buy or sell his/her used car gets confused with the price listed on the ads or in local market. They usually confirm the price from the market professionals which is a waste of time and effort. The system can be effective in filling such gaps which enables the users to predict the price of vehicles according to market value.